fix: fixed issue with health slider not showing correct value

+++ Assets/Scripts/Game/UI/WizardUI.cs
@@ -22,7 +22,7 @@ public class WizardUI : MonoBehaviour
     {
         IWizard wizard = GetWizard();
 
-        float ratio = wizard.Health / wizard.MaxHealth;
+        float ratio = wizard.Health / (float) wizard.MaxHealth;
         healthSlider.value = ratio;
     }
